test: complete CLI target and current-schema fixture migration
Move prepared target types to their canonical contract module to remove a type-only cycle. Build node-placement fixture targets from the canonical resolver and preserve admitted session input. Use the current schema constant in existing same-version eligibility fixtures without changing runtime schemas or assertions.
